Refactor logic to module
