+++ handlers/user-benefits/src/index.ts
@@ -1,5 +1,10 @@
 import { ValidationError } from '@modules/errors';
 import { IdentityApiGatewayAuthenticator } from '@modules/identity/apiGateway';
+import type { UserBenefitsResponse } from '@modules/product-benefits/schemas';
+import {
+	getBenefits,
+	getUserProducts,
+} from '@modules/product-benefits/userBenefits';
 import { getProductCatalogFromApi } from '@modules/product-catalog/api';
 import { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import type { Stage } from '@modules/stage';
@@ -8,10 +13,31 @@ import type {
 	APIGatewayProxyResult,
 	Handler,
 } from 'aws-lambda';
-import { getUserBenefits } from './userBenefits';
+import { getTrialInformation } from './trials';
 
 const stage = process.env.STAGE as Stage;
 const identityAuthenticator = new IdentityApiGatewayAuthenticator(stage, []); //TODO: Do we have any required scopes?
+
+const getUserBenefitsResponse = async (
+	stage: Stage,
+	productCatalogHelper: ProductCatalogHelper,
+	identityId: string,
+): Promise<UserBenefitsResponse> => {
+	const userProducts = await getUserProducts(
+		stage,
+		productCatalogHelper,
+		identityId,
+	);
+	console.log(`User products for user ${identityId} are: `, userProducts);
+	const benefits = getBenefits(userProducts);
+	console.log(`Benefits for user ${identityId} are: `, benefits);
+	const trials = getTrialInformation(benefits);
+	console.log(`Trials for user ${identityId} are: `, trials);
+	return {
+		benefits,
+		trials,
+	};
+};
 
 export const handler: Handler = async (
 	event: APIGatewayProxyEvent,
@@ -31,7 +57,7 @@ export const handler: Handler = async (
 		console.log(`Identity ID is ${authenticatedEvent.identityId}`);
 		const productCatalog = await getProductCatalogFromApi(stage);
 
-		const userBenefitsResponse = await getUserBenefits(
+		const userBenefitsResponse = await getUserBenefitsResponse(
 			stage,
 			new ProductCatalogHelper(productCatalog),
 			authenticatedEvent.identityId,
