+++ modules/product-benefits/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "scripts": {
     "build": "tsc --noEmit",
+    "test": "jest --group=-integration",
     "check-formatting": "prettier --check **.ts",
     "fix-formatting": "prettier --write **.ts"
   },
