--- modules/product-benefits/src/trialInformation.ts
@@ -1,12 +1,0 @@
-import type { Trial } from './schemas';
-
-// This information is used by the apps to work out what offer to show to the user
-export const feastExtendedTrial: Trial = {
-	iosSubscriptionGroup: '21445388',
-	androidOfferTag: 'initial_supporter_launch_offer',
-};
-
-export const feastRegularSubscription: Trial = {
-	iosSubscriptionGroup: '21396030',
-	androidOfferTag: 'initial_supporter_launch_offer',
-};
